requirements.txt
asgiref==3.2.10
black==21.11b1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.1
django-environ==0.8.1
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
huey==2.4.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.2
pyparsing==3.0.6
pytz==2021.3
redis==4.0.2
regex==2021.11.10
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tomli==1.2.2
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
